begin implementation new database structure
